Renamed RemoveVariable action namespace to DeleteVariable
