+++ src/Take.Blip.Builder/Hosting/ContainerExtensions.cs
@@ -7,6 +7,7 @@ using StackExchange.Redis;
 using System;
 using Take.Blip.Builder.Actions;
 using Take.Blip.Builder.Actions.CreateTicket;
+using Take.Blip.Builder.Actions.DeleteVariable;
 using Take.Blip.Builder.Actions.ExecuteScript;
 using Take.Blip.Builder.Actions.ForwardMessageToDesk;
 using Take.Blip.Builder.Actions.ManageList;
@@ -14,7 +15,6 @@ using Take.Blip.Builder.Actions.MergeContact;
 using Take.Blip.Builder.Actions.ProcessCommand;
 using Take.Blip.Builder.Actions.ProcessHttp;
 using Take.Blip.Builder.Actions.Redirect;
-using Take.Blip.Builder.Actions.RemoveVariable;
 using Take.Blip.Builder.Actions.SendCommand;
 using Take.Blip.Builder.Actions.SendMessage;
 using Take.Blip.Builder.Actions.SendMessageFromHttp;
