+++ src/Take.Blip.Client/Take.Blip.Client.csproj
@@ -17,9 +17,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Lime.Messaging" Version="0.8.107-beta" />
-    <PackageReference Include="Lime.Protocol" Version="0.8.107-beta" />
-    <PackageReference Include="Lime.Transport.Tcp" Version="0.8.107-beta" />
+    <PackageReference Include="Lime.Messaging" Version="0.8.107" />
+    <PackageReference Include="Lime.Protocol" Version="0.8.107" />
+    <PackageReference Include="Lime.Transport.Tcp" Version="0.8.107" />
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="2.0.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="2.0.0" />
     <PackageReference Include="Serilog" Version="2.6.0" />
